+++ src/dependencies/rmd_driver/src/driver.cpp
@@ -16,30 +16,30 @@ namespace rmd_driver{
     }
 
 
-    void Driver::addMotor(std::uint32_t actuator_id){
+    void Driver::addMotor(std::uint8_t actuator_id){
         updateIds(actuator_id);
     }
 
-    void Driver::MotorRunning(std::uint32_t actuator_id){
+    void Driver::MotorRunning(std::uint8_t actuator_id){
         MotorRunningRequest const request {};
         auto const response {sendRecv<MotorRunningResponse>(actuator_id,request)};
         return;
     }
 
 
-    Feedback Driver::sendTorqueSetpoint(std::uint32_t actuator_id,float const current) {
+    Feedback Driver::sendTorqueSetpoint(std::uint8_t actuator_id,float const current) {
         SetTorqueRequest const request {current};
         auto const response {sendRecv<SetTorqueResponse>(actuator_id,request)};
         return response.getStatus();
     }
 
-    void Driver::shutdownMotor(std::uint32_t actuator_id) {
+    void Driver::shutdownMotor(std::uint8_t actuator_id) {
         ShutdownMotorRequest const request {};
         auto const response {sendRecv<ShutdownMotorResponse>(actuator_id,request)};
         return;
     }
 
-    void Driver::stopMotor(std::uint32_t actuator_id) {
+    void Driver::stopMotor(std::uint8_t actuator_id) {
         StopMotorRequest const request {};
         auto const response {sendRecv<StopMotorResponse>(actuator_id,request)};
         return;
